Automatic File Extension Handling
If user does not specify a file extension when saving a conversation, it defaults to adding .txt

If when loading a conversation, a user does not enter a file extension and the file is not found, it will try looking for files with .txt, then it will check any with that base name and any file extension. If multiple matches are found displays an error.